# services/gateway/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EcommerceGo API Gateway",
    "description": "Unified entry point for all EcommerceGo microservices. All requests require a JWT Bearer token in the Authorization header unless noted otherwise. Prices are always in cents (int64).",
    "version": "1.0.0",
    "contact": {
      "name": "EcommerceGo",
      "url": "https://github.com/utafrali/EcommerceGo"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development (API Gateway)"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register new user",
        "operationId": "register",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } } },
        "responses": {
          "201": { "description": "User registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "description": "Email already exists" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login",
        "operationId": "login",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
        "responses": {
          "200": { "description": "Login successful", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Refresh access token",
        "operationId": "refreshToken",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } } },
        "responses": {
          "200": { "description": "Token refreshed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login or /api/v1/auth/register"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Missing or invalid JWT token",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    },
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "UNAUTHORIZED" },
          "message": { "type": "string", "example": "missing or invalid token" },
          "request_id": { "type": "string", "example": "req-abc123" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "first_name", "last_name"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refresh_token"],
        "properties": {
          "refresh_token": { "type": "string" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "access_token": { "type": "string" },
              "refresh_token": { "type": "string" },
              "expires_in": { "type": "integer", "description": "Token expiry in seconds" },
              "user_id": { "type": "string", "format": "uuid" }
            }
          }
        }
      },
...
    }
  }
}